requirements.txt
google-genai>=0.5.0
chromadb>=0.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
fastapi>=0.110.0
